Refactor reducer
action-handler 삭제
action 처리 관련 코드를 reducer.js에서 관리
